requirements.txt
playwright
pyyaml
ccxt
hypothesis
pytest
pytest-asyncio
openai
